+++ packages/terra-profile-image/tests/jest/ProfileImage.test.jsx
@@ -6,3 +6,9 @@ it('should render a Terra image with a placeholder prop', () => {
   const wrapper = shallow(<ProfileImage src="profile.jpg" height="75" width="75" alt="A profile image" />);
   expect(wrapper).toMatchSnapshot();
 });
+
+/* Profile Image source is not provided */
+it('should render a Terra image without Source prop', () => {
+  const wrapper = shallow(<ProfileImage src="" height="75" width="75" alt="Source not provided" />);
+  expect(wrapper).toMatchSnapshot();
+});